Permissions card is now a separate component that can be imported.
